add GetDocumentExtender to interface
